refactor: move tool discovery logic into dedicated tool_loader module
